Changed and compiled contracts and added new tests

+++ src/ethereum/contracts/FootballGameBet.sol
@@ -2,9 +2,9 @@ pragma solidity  >=0.4.22 <0.7.0;
 
 contract FootballGameBet {
     string public homeTeam;
-    uint8 public homeTeamGoals;
+    uint public homeTeamGoals;
     string public awayTeam;
-    uint8 public awayTeamGoals;
+    uint public awayTeamGoals;
     uint public startTime;
     
     address payable public organiser;
@@ -46,11 +46,11 @@ contract FootballGameBet {
         return players;
     }
     
-    function gameFinished(uint8 homeGoals, uint8 awayGoals) public isContractBalanceValid isOrganiser {
+    function gameFinished(uint homeGoals, uint awayGoals) public isContractBalanceValid isOrganiser {
         homeTeamGoals = homeGoals;
         awayTeamGoals = awayGoals;
         
-        uint8 winner = 0;
+        uint winner = 0;
         uint winningAmount = players.length*stake;
         uint organiserFee = 0;
